move redeploy button params parsing to application code
